Move menu & back press handling to the fragment + import cleanup
