+++ app/src/main/java/io/particle/android/sdk/ui/devicelist/DeviceListActivity.kt
@@ -2,29 +2,20 @@ package io.particle.android.sdk.ui.devicelist
 
 import android.content.Intent
 import android.os.Bundle
-import android.view.Menu
-import android.view.MenuItem
-import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.commit
-import io.particle.android.sdk.accountsetup.LoginActivity
-import io.particle.android.sdk.cloud.ParticleCloudSDK
 import io.particle.android.sdk.cloud.ParticleDevice
 import io.particle.android.sdk.ui.BaseActivity
-import io.particle.android.sdk.utils.Py.list
 import io.particle.android.sdk.utils.SoftAPConfigRemover
 import io.particle.android.sdk.utils.WifiFacade
 import io.particle.android.sdk.utils.ui.Ui
 import io.particle.mesh.ui.setup.MeshSetupActivity
 import io.particle.sdk.app.R
-import pl.brightinventions.slf4android.LogRecord
-import pl.brightinventions.slf4android.NotifyDeveloperDialogDisplayActivity
 
 
 class DeviceListActivity : BaseActivity(), DeviceListFragment.Callbacks {
 
     private var softAPConfigRemover: SoftAPConfigRemover? = null
-    private var deviceList: DeviceListFragment? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -32,7 +23,6 @@ class DeviceListActivity : BaseActivity(), DeviceListFragment.Callbacks {
 
         softAPConfigRemover = SoftAPConfigRemover(this, WifiFacade.get(this))
 
-        deviceList = Ui.findFrag(this, R.id.fragment_device_list)
         // TODO: If exposing deep links into your app, handle intents here.
 
         if (Ui.findFrag<Fragment>(this, R.id.fragment_parent) == null) {
@@ -68,47 +58,14 @@ class DeviceListActivity : BaseActivity(), DeviceListFragment.Callbacks {
     }
 
     override fun onBackPressed() {
-        if (deviceList?.onBackPressed() == false) {
+        val currentFragment = Ui.findFrag<Fragment>(this, R.id.fragment_parent)
+        var deviceList: DeviceListFragment? = null
+        if (currentFragment is DeviceListFragment) {
+            deviceList = currentFragment
+        }
+        if (deviceList?.onBackPressed() != true) {
             super.onBackPressed()
         }
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu): Boolean {
-        menuInflater.inflate(R.menu.device_list, menu)
-        return super.onCreateOptionsMenu(menu)
-    }
-
-    private fun sendLogs() {
-        val lr = LogRecord(java.util.logging.Level.WARNING, "")
-
-        NotifyDeveloperDialogDisplayActivity.showDialogIn(
-            this,
-            lr,
-            list(),
-            "Logs from the Particle Android app",
-            "",
-            list("pl.brightinventions.slf4android.ReadLogcatEntriesAsyncTask")
-        )
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        val id = item.itemId
-        when (id) {
-            R.id.action_log_out -> AlertDialog.Builder(this)
-                .setMessage(R.string.logout_confirm_message)
-                .setPositiveButton(R.string.log_out) { dialog, _ ->
-                    val cloud = ParticleCloudSDK.getCloud()
-                    cloud.logOut()
-                    startActivity(Intent(this@DeviceListActivity, LoginActivity::class.java))
-                    finish()
-
-                    dialog.dismiss()
-                }
-                .setNegativeButton(R.string.cancel) { dialog, _ -> dialog.dismiss() }
-                .show()
-            R.id.action_send_logs -> sendLogs()
-        }
-        return super.onOptionsItemSelected(item)
     }
 
     //region DeviceListFragment.Callbacks
